r2engine: rendering engine refactor - save point

+++ r2engine/src/r2/Render/Backends/SDL_OpenGL/OpenGLRingBuffer.h
@@ -10,8 +10,8 @@ namespace r2::draw::rendererimpl
 {
 	struct BufferRange
 	{
-		u64 startOffset;
-		u64 length;
+		u32 startOffset;
+		u32 length;
 	};
 
 	bool BufferRangeOverlaps(const BufferRange& _lhs, const BufferRange& _rhs);
